+++ Mlib/Osm_Loader/Osm_Map_Resource/Add_Trees_To_Forest_Outlines.hpp
@@ -14,13 +14,13 @@ void add_trees_to_forest_outlines(
     BatchResourceInstantiator& bri,
     // std::list<SteinerPointInfo>& steiner_points,
     ResourceNameCycle& rnc,
-    float min_dist_to_road,
+    double min_dist_to_road,
     const StreetBvh& street_bvh,
     const GroundBvh& ground_bvh,
     const std::map<std::string, Node>& nodes,
     const std::map<std::string, Way>& ways,
-    float tree_distance,
-    float tree_inwards_distance,
-    float scale);
+    double tree_distance,
+    double tree_inwards_distance,
+    double scale);
 
 }
