+++ Mlib/Osm_Loader/Osm_Map_Resource/Calculate_Spawn_Points.hpp
@@ -10,7 +10,7 @@ struct StreetRectangle;
 void calculate_spawn_points(
     std::list<SpawnPoint>& spawn_points,
     const std::list<StreetRectangle>& street_rectangles,
-    float scale,
+    double scale,
     DrivingDirection driving_direction);
 
 }
