+++ Mlib/Osm_Loader/Osm_Map_Resource/Extended_Image.hpp
@@ -6,15 +6,15 @@ namespace Mlib {
 class ExtendedImage {
 public:
     ExtendedImage(
-        const Array<float>& image,
+        const Array<double>& image,
         const Array<bool>& mask,
         size_t extension,
         size_t box_filter_radius,
         size_t niterations,
         bool preserve_original = true);
-    bool operator () (float r, float c, float& value) const;
+    bool operator () (double r, double c, double& value) const;
 private:
-    Array<float> extended_image_;
+    Array<double> extended_image_;
     size_t extension_;
 };
