+++ Mlib/Osm_Loader/Osm_Map_Resource/Get_Map_Outer_Contour.hpp
@@ -10,7 +10,7 @@ class FixedArray;
 struct Node;
 struct Way;
 
-std::vector<FixedArray<float, 2>> get_map_outer_contour(
+std::vector<FixedArray<double, 2>> get_map_outer_contour(
     const std::map<std::string, Node>& nodes,
     const std::map<std::string, Way>& ways);
