+++ Mlib/Osm_Loader/Osm_Map_Resource/Get_Smooth_Way.hpp
@@ -9,10 +9,10 @@ struct Node;
 template <typename TData, size_t... tshape>
 class FixedArray;
 
-std::list<FixedArray<float, 2>> smooth_way(
+std::list<FixedArray<double, 2>> smooth_way(
     const std::map<std::string, Node>& nodes,
     const std::list<std::string>& nd,
-    float scale,
-    float max_length);
+    double scale,
+    double max_length);
 
 }
