+++ Mlib/Osm_Loader/Osm_Map_Resource/Get_Terrain_Region_Contours.cpp
@@ -6,11 +6,11 @@
 
 using namespace Mlib;
 
-std::list<std::pair<TerrainType, std::list<FixedArray<float, 3>>>> Mlib::get_terrain_region_contours(
+std::list<std::pair<TerrainType, std::list<FixedArray<double, 3>>>> Mlib::get_terrain_region_contours(
     const std::map<std::string, Node>& nodes,
     const std::map<std::string, Way>& ways)
 {
-    std::list<std::pair<TerrainType, std::list<FixedArray<float, 3>>>> result;
+    std::list<std::pair<TerrainType, std::list<FixedArray<double, 3>>>> result;
     for (const auto& w : ways) {
         const auto& tags = w.second.tags;
         if (tags.contains("layer") &&
