+++ Mlib/Osm_Loader/Osm_Map_Resource/Get_Water_Region_Contours.hpp
@@ -11,7 +11,7 @@ enum class WaterType;
 template <typename TData, size_t... tshape>
 class FixedArray;
 
-std::list<std::pair<WaterType, std::list<FixedArray<float, 3>>>> get_water_region_contours(
+std::list<std::pair<WaterType, std::list<FixedArray<double, 3>>>> get_water_region_contours(
     const std::map<std::string, Node>& nodes,
     const std::map<std::string, Way>& ways);
